poc/requirements.txt
# Sarvam AI SDK
sarvamai>=0.1.0

# UI
streamlit>=1.32.0
audio-recorder-streamlit>=0.0.10

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.24.0

# Embeddings
sentence-transformers>=2.2.2

# PDF parsing
pdfplumber>=0.10.0

# Audio I/O
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
Authlib>=1.3.2
emails
